+++ FE/src/pages/MyPage.tsx
@@ -13,6 +13,7 @@ import { useAppDispatch, useAppSelector } from "@hooks/hook";
 import { selectUserId } from "@store/authSlice";
 import { setCategoryColors, setCategoryInput, setCategoryList, setDdayList } from "@store/mypageSlice";
 import { CategoryConfig } from "@util/planner.interface";
+import CancleMembership from "@components/mypage/details/myInfo/CancleMembership";
 
 const MyPage = () => {
   const dispatch = useAppDispatch();
@@ -72,6 +73,7 @@ const MyPage = () => {
             {
               "내 정보 확인": <MyPageInfo />,
               "비밀번호 변경": <MyPassword />,
+              "회원 탈퇴": <CancleMembership />,
               "다이어리 설정": <MyPageDiary />,
               "카테고리 설정": <MyPageFrame title="카테고리" />,
               "디데이 설정": <MyPageFrame title="디데이" />,
